backend/requirements.txt
# MindSync: Multimodal AI Framework for Mental Health Monitoring
# Verified against Python 3.11 — uses >= to avoid downgrading already-installed packages

# Core Deep Learning
torch>=2.0.1
torchaudio>=2.0.2

# HuggingFace Ecosystem
transformers>=4.30.2
datasets>=2.13.1
tokenizers>=0.13.3
accelerate>=0.20.3
evaluate>=0.4.0
huggingface-hub>=0.15.1

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1

# Scientific Computing
numpy>=1.24.3
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.3

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.15.0
Pillow>=10.0.0

# NLP Utilities
nltk>=3.8.1
ftfy>=6.1.1

# Training Utilities
tqdm>=4.65.0
wandb>=0.15.5

# Configuration
PyYAML>=6.0
omegaconf>=2.3.0

# Demo App
gradio>=3.39.0

# Testing
pytest>=7.4.0

# Statistical Analysis
statsmodels>=0.14.0
